+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # Changes in v0.3 (not released)
 - Add support for VolumeStep and VolumeType.UpDown
 - Add support for output configuration
+- Add support for PlayOrPause command
 
 # Changes in v0.2 (released 2025-02-14)
 - First official release