+++ ChangeLog.md
@@ -2,6 +2,7 @@
 - Add support for UpDown volume type and new volume commands
 - Add support for output device configuration
 - Add support for PlayOrPause command
+- Add support for API permissions
 
 # Changes in v0.2 (released 2025-02-14)
 - First official release