+++ ChangeLog.md
@@ -3,6 +3,7 @@
 - Add support for output device configuration
 - Add support for PlayOrPause command
 - Add support for API permissions
+- Add support for HTTP basic auth
 
 # Changes in v0.2 (released 2025-02-14)
 - First official release